% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cindex.R
\name{Cindex}
\alias{Cindex}
\title{Compute the Cindex}
\usage{
Cindex(mod, newdat0, digits = 3)
}
\arguments{
\item{digits}{to round the Cindex; if digits=NULL, do not round
Produce metrics to evaluate a survival model}
}
\description{
Compute the Cindex
}
