% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineEventDate.R
\name{defineEventDate}
\alias{defineEventDate}
\title{Define event date}
\usage{
defineEventDate(ed, cutoff, event = "OS", format.dates.in.ed = TRUE,
  cutoff.date.format = "\%Y-\%m-\%d", status.only = FALSE)
}
\arguments{
\item{ed}{input data, typically a data.frame or matrix}

\item{cutoff}{cutoff date}

\item{event}{event type; one of OS(default), DSS, or RFS}

\item{format.dates.in.ed}{logical; if \code{TRUE} (default), dates will be
formatted in \code{ed}}

\item{cutoff.date.format}{format of cutoff date}

\item{status.only}{logical; if \code{TRUE}, only survival status will be returned}
}
\value{
A list with elements
\item{ev.status}{A string indicating status: such as "os.censor" or "os.event}
\item{ev.Date}{Date object}
\item{ev.years}{Date in years}
}
\description{
Calculate event/status given censor date
}
\details{
Defines the event dates given a vector of cutoff dates. If the cutoff is a
single value, it is repeated to equal the number of rows of \code{ed}.
}
\note{
Since there may be some string, e.g. "Unk" in the return date
\code{ev.Date}, R changed the date to string. Therefore, you need to do
\code{as.numeric()} and \code{as.Date()} to change \code{ev.Date} from
a string back to a \code{Date} object.
}
\examples{
library(bccaEndometrial)
data(ed)
Year <- as.numeric(format(as.Date(ed$followup.start.date.mm.dd.yyyy, format = "\%m/\%d/\%Y"), "\%Y"))
cutoffDate <- sapply(Year + 5, function(x) paste0("12/31/", x))
defineEventDate(ed, cutoffDate, cutoff.date.format = "\%m/\%d/\%Y", event = "RFS")
defineEventDate(ed, cutoffDate, cutoff.date.format = "\%m/\%d/\%Y", event = "DSS")
}
\author{
Samuel Leung, Aline Talhouk
}

