% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genBootSpls.R
\name{genBootSpls}
\alias{genBootSpls}
\title{Standard deviation of log hazard ratio}
\usage{
genBootSpls(dat, B = 1000)
}
\arguments{
\item{dat, }{the data matrix that needs to be bootstrapped}

\item{B, }{the number of bootstrap samples desired (default to 1000)}
}
\value{
a list with with boot.tr being the bootstrap samples and boot.te being the unseleceted samples
}
\description{
Compute the standard deviation of a log hazard ratio when only given its
repoted confidence limits.
}
\details{
Extract standard deviation from a confidence limit and obtain a
corresponding p-value, which can be compared to the reported p-value.
The null p-values are calculated from the Chi-squared 1 distribution.
Generating Bootstrap Samples of a dataset

Generate bootstrap samples by sampling with replacement. The bootstrap samples and the
unselected samples are saved seperately.
}
\author{
Aline Talhouk, Derek Chiu
}

