% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCox.R
\name{predictCox}
\alias{predictCox}
\title{A function to make prediction in cox model (works for both coxph and coxphf)}
\usage{
predictCox(mod, newdata)
}
\description{
A function to make prediction in cox model (works for both coxph and coxphf)
}

