% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPerf.R
\name{bootPerf}
\alias{bootPerf}
\title{Assess Performance}
\usage{
bootPerf(dat0, var.names, time.var, event.ind, B = 1000, seed = 2014,
  firth = TRUE)
}
\description{
Produce metrics to evaluate a survival model in Bootstrap samples
}

