% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPerf.R
\name{bootPerf}
\alias{bootPerf}
\title{Assess Performance of a survival model}
\usage{
bootPerf(dat0, var.names, time.var, event.ind, B = 1000, seed = 2014,
  firth = TRUE)
}
\arguments{
\item{dat0}{is the training data}

\item{var.names}{are the names of the predictors}

\item{event.ind}{is the event indicator}

\item{B}{is the number of bootstrap iterations defaults to 1000}

\item{seed}{is a random number generator seed set to 2014}

\item{firth}{whether or not to perform a firth correction set to TRUE}

\item{is}{the time variable}
}
\value{
\item{apparent}{the apparent fit: testing the model on the full training
data}
\item{boot}{the bootstrap fit: testing the model on the bootstrapped training
data}
\item{boot632}{the bootstrap 0.632 fit: testing the model on the bootstrapped
test data, ie the data that was not selected for bootstrap training}
}
\description{
Produce metrics to evaluate a survival model in Bootstrap samples (Internal
Validation)
}
\author{
Aline Talhouk
}

