% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposeSurvform.R
\name{decomposeSurvform}
\alias{decomposeSurvform}
\title{a function to decompose a survival formula}
\usage{
decomposeSurvform(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data}
}
\description{
a function to decompose a survival formula
}

