% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCox.R
\name{doCox}
\alias{doCox}
\title{Perform Coxph or Coxphf}
\usage{
doCox(formula, data0, firth = TRUE)
}
\arguments{
\item{formula}{formula}

\item{data0}{data object}

\item{firth}{logical; if \code{TRUE}, uses firth correction}
}
\description{
Perform Coxph or Coxphf
}

